+++ sampledb/logic/background_tasks/__init__.py
@@ -2,6 +2,7 @@
 from . import core
 from . import send_mail
 from . import background_dataverse_export
+from .automatic_schema_updates import post_check_for_automatic_schema_updates_task, post_perform_automatic_schema_updates_task
 from .core import start_handler_threads, stop_handler_threads, post_background_task, get_background_tasks, get_background_task_result, reset_claimed_background_tasks
 from .send_mail import post_send_mail_task
 from .background_dataverse_export import post_dataverse_export_task
@@ -18,7 +19,9 @@ __all__ = [
     'post_background_task',
     'get_background_tasks',
     'get_background_task_result',
+    'post_check_for_automatic_schema_updates_task',
     'post_dataverse_export_task',
+    'post_perform_automatic_schema_updates_task',
     'post_poke_components_task',
     'post_trigger_object_log_webhooks',
     'reset_claimed_background_tasks',